requirements.txt
achallonge-compat-fork==1.9.0
adrf==0.1.9
aenum==3.1.15
aiofiles==24.1.0
aiohappyeyeballs==2.4.4
aiohttp==3.9.5
aiosignal==1.3.1
alembic==1.14.0
amqp==5.2.0
annotated-types==0.7.0
anyio==4.4.0
asgiref==3.8.1
asn1crypto==1.5.1
asttokens==2.4.1
async-property==0.2.2
async-timeout==4.0.3
asyncio==3.4.3
attrs==23.2.0
base58==2.1.1
beautifulsoup4==4.13.3
bidict==0.23.1
billiard==4.2.0
bip32==3.4
bip39==0.0.2
bip44==0.1.4
bitarray==2.9.2
blinker==1.9.0
cachetools==5.5.1
celery==5.4.0
certifi==2024.6.2
cffi==1.16.0
charset-normalizer==3.3.2
ckzg==2.0.0
click==8.1.7
click-didyoumean==0.3.1
click-plugins==1.1.1
click-repl==0.3.0
coincurve==18.0.0
colorama==0.4.6
comm==0.1.4
construct==2.10.68
construct-typing==0.5.6
cron-descriptor==1.4.3
cryptography==41.0.7
cytoolz==0.12.3
dataclasses-json==0.6.7
debugpy==1.8.0
decorator==5.1.1
deepgram-sdk==2.12.0
deprecation==2.1.0
diff-match-patch==20230430
discord==2.3.2
discord.py==2.4.0
distro==1.9.0
Django==5.1
django-celery-beat==2.7.0
django-celery-results==2.5.1
django-cors-headers==4.4.0
django-debug-toolbar==4.4.6
django-import-export==4.1.1
django-ratelimit==4.1.0
django-timezone-field==7.0
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
eth-account==0.13.3
eth-hash==0.7.0
eth-keyfile==0.8.1
eth-keys==0.5.1
eth-rlp==2.1.0
eth-typing==5.0.0
eth-utils==5.0.0
eth_abi==5.1.0
exceptiongroup==1.2.2
executing==2.0.1
Flask==3.1.0
Flask-Cors==5.0.0
Flask-JWT-Extended==4.7.1
Flask-Migrate==4.1.0
Flask-Neo4j==0.5.1
Flask-SQLAlchemy==3.1.1
frappe-bench==5.23.0
frozendict==2.3.10
frozenlist==1.4.1
future==1.0.0
gitdb==4.0.12
GitPython==3.1.44
google-ai-generativelanguage==0.6.15
google-api-core==2.24.2
google-api-python-client==2.165.0
google-auth==2.38.0
google-auth-httplib2==0.2.0
google-generativeai==0.8.4
googleapis-common-protos==1.69.2
greenlet==3.1.1
grpcio==1.71.0
grpcio-status==1.71.0
grpcio-tools==1.64.1
gunicorn==22.0.0
h11==0.14.0
hexbytes==1.2.1
honcho==2.0.0
httpcore==1.0.5
httplib2==0.22.0
httpx==0.27.0
idna==3.7
imageio==2.36.1
imageio-ffmpeg==0.5.1
iniconfig==2.0.0
interchange==2021.0.4
ipykernel==6.26.0
ipython==8.17.2
itsdangerous==2.2.0
jedi==0.19.1
Jinja2==3.1.4
jiter==0.8.2
joblib==1.4.2
jsonalias==0.1.1
jsonschema==4.22.0
jsonschema-specifications==2023.12.1
jupyter_client==8.5.0
jupyter_core==5.5.0
kombu==5.4.0
loguru==0.7.0
lru-dict==1.2.0
Mako==1.3.8
MarkupSafe==3.0.2
marshmallow==3.23.1
matplotlib-inline==0.1.6
mnemonic==0.20
mock==5.1.0
monotonic==1.6
moralis==0.1.49
moviepy==2.1.1
multidict==6.0.5
multitasking==0.0.11
mypy-extensions==1.0.0
neo4j==5.27.0
nest-asyncio==1.6.0
nltk==3.9.1
numpy==2.2.0
oauthlib==3.2.2
openai==1.57.4
packaging==24.1
pandas==2.2.3
pansi==2024.11.0
parsimonious==0.10.0
parso==0.8.3
peewee==3.17.9
Pillow==10.0.0
platformdirs==3.11.0
pluggy==1.5.0
proglog==0.1.10
prompt_toolkit==3.0.47
propcache==0.2.1
proto-plus==1.26.1
protobuf==5.27.2
psutil==5.9.6
psycopg==3.2.4
psycopg-binary==3.2.4
psycopg-pool==3.2.4
psycopg2-binary==2.9.10
pure-eval==0.2.2
py2neo==2021.2.4
pyasn1==0.6.1
pyasn1_modules==0.4.1
pybip39==0.1.0
pycparser==2.22
pycryptodome==3.20.0
pydantic==2.8.2
pydantic_core==2.20.1
Pygments==2.16.1
PyJWT==2.9.0
pylightning==0.0.7.3
pyparsing==3.2.1
pytest==8.3.4
python-crontab==2.6.0
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-engineio==4.5.1
python-socketio==5.8.0
pytz==2024.2
pyunormalize==15.1.0
pywin32==306
pyzmq==25.1.1
redis==5.0.7
referencing==0.35.1
regex==2024.5.15
requests==2.32.3
requests-oauthlib==2.0.0
rlp==4.0.1
rpds-py==0.18.1
rsa==4.9
semantic-version==2.10.0
six==1.16.0
smmap==5.0.2
sniffio==1.3.1
solana==0.34.0
solders==0.21.0
soupsieve==2.6
split==0.4
SQLAlchemy==2.0.36
sqlparse==0.5.0
stack-data==0.6.3
stripe==5.0.0
tablib==3.5.0
textblob==0.18.0.post0
toolz==0.12.1
tornado==6.3.3
tqdm==4.67.1
traitlets==5.13.0
twitter==1.19.6
types-requests==2.32.0.20240712
typing-inspect==0.9.0
typing_extensions==4.12.2
tzdata==2024.1
uritemplate==4.1.1
urllib3==2.2.2
uvicorn==0.34.0
vine==5.1.0
wcwidth==0.2.13
web3==7.2.0
websockets==11.0.3
Werkzeug==3.1.3
whitenoise==6.7.0
win32_setctime==1.2.0
yarl==1.9.4
yfinance==0.2.55
Flask
requests
yfinance
openai
gemini  # If using Google AI APIs
